pyproject.toml
[tool.poetry]
name = "lynbrook-app-backend"
version = "0.1.0"
description = ""
authors = ["Oliver Ni <oliver.ni@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
psycopg2 = "^2.9.1"
django-stubs = "^1.8.0"
django-dotenv = "^1.4.2"
django-better-admin-arrayfield = "^1.4.2"
rules = "^3.0"
django-allauth = "^0.45.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
ipython = "^7.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
